requirements.txt
streamlit
openai
python-docx
python-pptx